pyproject.toml
[project]
name = "scripts_de_python"
version = "1"
description = ""
authors = [
    {name = "jonathan correa",email = "jonathancorrea291292@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "easyocr (>=1.7.2,<2.0.0)",
    "pymupdf (>=1.25.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
